New Dev command working (logger WIP)
* The getAvailablePort function has been added to leemons-utils (get the
next free port)

* Removed Dev watchers per service, replaced by a ReloaderProvider (DRY!)
  * A reload timeout is added to wait for more changes
  * When new changes occurs, the timeout starts again.
  * Wait current reload to finish

* Added process.title for showing a custom process name
  * Master: 'Leemons Dev'
  * Worker: 'Leemons Dev Instance'

* Added new conditions to next.js building:
  * Do not build if is a 'development' environment
  * Rebuild if next/.next/BUILD_ID does not exists

* Added a createWorker function (DRY!)

* Added a stdin handler
  * When typed "rs", "restart" or "reload" + enter, the leemons app is
    reloaded

  * When pressed ctrl+l+enter the console is cleared

+++ packages/leemons/lib/core/front/build.js
@@ -38,30 +38,33 @@ async function frontDeps() {
  * Next.js Build
  */
 async function buildNext() {
-  // Check if the build exists
-  if (!(await fs.exists(path.resolve(leemons.dir.next, '.next')))) {
-    leemons.frontNeedsBuild = true;
-    // The error can be related to dependencies
-    leemons.frontNeedsUpdateDeps = true;
-  }
+  // Do not build if is a dev environment
+  if (process.env.NODE_ENV !== 'development') {
+    // Check if a production build exists
+    if (!(await fs.exists(path.resolve(leemons.dir.next, '.next', 'BUILD_ID')))) {
+      leemons.frontNeedsBuild = true;
+      // The error can be related to dependencies
+      leemons.frontNeedsUpdateDeps = true;
+    }
 
-  // Build if necessary
-  if (leemons.frontNeedsBuild && process.env.NODE_ENV !== 'development') {
-    await hooks.fireEvent('leemons::build', { status: 'start' });
-    const spinner = ora('Building frontend').start();
-    return execa
-      .command(`yarn --cwd ${leemons.dir.next} build`)
-      .then(() => {
-        spinner.succeed('Frontend builded');
-      })
-      .catch(async (e) => {
-        spinner.fail(`Frontend can't be builded`);
-        await fs.remove(path.resolve(leemons.dir.next, '.next'));
-        throw e;
-      })
-      .finally(async () => {
-        await hooks.fireEvent('leemons::build', { status: 'end' });
-      });
+    // Build if necessary
+    if (leemons.frontNeedsBuild && process.env.NODE_ENV !== 'development') {
+      await hooks.fireEvent('leemons::build', { status: 'start' });
+      const spinner = ora('Building frontend').start();
+      return execa
+        .command(`yarn --cwd ${leemons.dir.next} build`)
+        .then(() => {
+          spinner.succeed('Frontend builded');
+        })
+        .catch(async (e) => {
+          spinner.fail(`Frontend can't be builded`);
+          await fs.remove(path.resolve(leemons.dir.next, '.next'));
+          throw e;
+        })
+        .finally(async () => {
+          await hooks.fireEvent('leemons::build', { status: 'end' });
+        });
+    }
   }
   return null;
 }
